pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "itd-sdk"
version = "2.0.0a1"
description = "ITD client for python"
readme = "README.md"
authors = [
  { name = "firedotguy", email = "nta16022013@gmail.com" }
]
license = "MIT"
dependencies = [
  "requests~=2.32.5",
  "pydantic~=2.12.5",
  "sseclient-py~=1.9.0",
  "websocket-client>=1.9.0",
]
requires-python = ">=3.9"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
]
